requirements.txt
# Firebase Cloud Functions Dependencies
firebase-functions>=0.1.0
firebase-admin>=6.2.0

# AI/ML APIs
openai>=1.3.0
elevenlabs>=0.2.0

# News and Content APIs
google-search-results>=2.4.2
feedparser>=6.0.10
gnews

# Web scraping and content processing
newspaper3k==0.2.8
lxml==5.4.0
lxml_html_clean==0.4.2
nltk==3.9.1
requests==2.32.3
urllib3==2.4.0

# Utility libraries
pathlib2>=2.3.7

# Optional: For development and testing
python-dotenv>=1.0.0

# Added from the code block
functions-framework==3.*
google-cloud-firestore
google-cloud-logging 